app_heart/requirements.txt
cloudpickle==2.0.0
numpy==1.21.5
pandas==1.3.5
pyjanitor==0.22.0
streamlit==1.22.0
streamlit_echarts==0.4.0
scikit-learn==1.0.2
xgboost==1.5.0
